Changed uint8_t to unsigned int.

+++ examples/test.c
@@ -24,7 +24,7 @@ static void chill(int speed) {
 }
 
 void test_led_on_off(monome_t *monome) {
-	uint8_t i, j;
+	unsigned int i, j;
 	
 	for( i = 0; i < 16; i++ )
 		for( j = 0; j < 16; j++ ) {
@@ -40,7 +40,7 @@ void test_led_on_off(monome_t *monome) {
 }
 
 void test_led_row(monome_t *monome) {
-	uint8_t i, buf_off[2], buf_on[2];
+	unsigned int i, buf_off[2], buf_on[2];
 	
 	buf_off[0] = buf_off[1] = 0x00;
 	buf_on[0]  = buf_on[1]  = 0xFF;
@@ -53,7 +53,7 @@ void test_led_row(monome_t *monome) {
 }
 
 void test_led_col(monome_t *monome) {
-	uint8_t i, buf_off[2], buf_on[2];
+	unsigned int i, buf_off[2], buf_on[2];
 	
 	buf_off[0] = buf_off[1] = 0x00;
 	buf_on[0]  = buf_on[1]  = 0xFF;
@@ -66,7 +66,7 @@ void test_led_col(monome_t *monome) {
 }
 
 void test_led_frame(monome_t *monome) {
-	uint8_t pat[2], buf[8], i, j, k, l;
+	unsigned int pat[2], buf[8], i, j, k, l;
 	
 	pat[0] = 0x55;
 	pat[1] = 0xAA;
@@ -87,7 +87,7 @@ void test_led_frame(monome_t *monome) {
 }
 
 void fade_out(monome_t *monome) {
-	uint8_t i = 0x10;
+	unsigned int i = 0x10;
 	
 	while( i-- ) {
 		monome_intensity(monome, i);
